Add DragonFly BSD support and improve FreeBSD support
Support DragonFly BSD for TUN creation and destroy in the same way as
FreeBSD.

No need to load the `if_tun` kernel module, because the `ifconfig(8)`
utility on FreeBSD and DragonFly BSD will auto load it if necessary
(the `-n` option suppresses this behavior; see man page for details).
This feature was added to `ifconfig(8)` in 1999:
https://github.com/freebsd/freebsd/commit/4d16916f800d4c5ee37fb3c93ca6981ad9b33eb3

In addition, FreeBSD 13 has merged `if_tun` and `if_tap` into a single
module named `if_tuntap`.  So this change improves the FreeBSD
support.

Signed-off-by: Aaron LI <aly@aaronly.me>
Signed-off-by: Daniel Lenski <dlenski@gmail.com>